+++ packages/base/CHANGELOG.md
@@ -6,7 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-
+ - `ServerRepository.selectAsArrayOfValues()` - for Repository with ONE attribute returns a flat array of attribute values
+   ```javascript
+    const usersIDs = UB.Repository('uba_user').attrs('ID').limit(100).selectAsArrayOfValues()
+    // usersIDs is array of IDs [1, 2, 3, 4]
+   ```
+   
 ### Changed
 
 ### Deprecated