+++ packages/cs-shared/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `CustomRepository.selectAsArrayOfValues()` - for Repository with ONE attribute returns a flat array of attribute values
 
 ### Changed